SSZ Annotations and auto-serialize: code style fixes
