+++ ssz/src/main/java/net/consensys/cava/ssz/SSZScheme.java
@@ -4,22 +4,26 @@ import java.util.ArrayList;
 import java.util.List;
 
 /**
- * Object SSZ scheme.
+ * <p>Object SSZ scheme.</p>
  *
- * Enumerates all object fields and their
- * properties in appropriate order.
+ * <p>Enumerates all object fields and their
+ * properties in appropriate order.</p>
  */
 public class SSZScheme {
-    List<SSZField> fields = new ArrayList<>();
+  List<SSZField> fields = new ArrayList<>();
 
-    static class SSZField {
-        Class type;
-        SSZSerializer.SSZType sszType;
-        String name;
-        String getter;
-        // container not needed : null
-        // container needed : false
-        // container needed but should be omitted: true
-        Boolean skipContainer = null;
-    }
+  static class SSZField {
+    Class type;
+    SSZSerializer.SSZType sszType;
+    String name;
+    String getter;
+    /**
+     * <ul>
+     * <li>Container not needed (primitive type) : null</li>
+     * <li>Container needed : false</li>
+     * <li>Container needed but should be omitted: true</li>
+     * </ul>
+     */
+    Boolean skipContainer = null;
+  }
 }
