+++ ssz/src/main/java/net/consensys/cava/ssz/annotation/SSZSerializable.java
@@ -1,18 +1,23 @@
 package net.consensys.cava.ssz.annotation;
 
-import java.lang.annotation.*;
+import java.lang.annotation.Documented;
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Inherited;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
 
 /**
- * Identifies class that could be SSZ serializable
+ * <p>Identifies class that could be SSZ serializable</p>
  */
 @Documented
 @Target(ElementType.TYPE)
 @Inherited
 @Retention(RetentionPolicy.RUNTIME)
 public @interface SSZSerializable {
-    /**
-     * If set, uses following method to get encoded class value
-     * @return method for encoding
-     */
-    String encode() default "";
+  /**
+   * <p>If set, uses following method to get encoded class data</p>
+   * @return method for encoding
+   */
+  String encode() default "";
 }
