+++ ssz/src/main/java/net/consensys/cava/ssz/annotation/SSZTransient.java
@@ -1,9 +1,14 @@
 package net.consensys.cava.ssz.annotation;
 
-import java.lang.annotation.*;
+import java.lang.annotation.Documented;
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Inherited;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
 
 /**
- * Skips field in SSZ representation
+ * <p>Skips field in SSZ representation</p>
  */
 @Documented
 @Target(ElementType.FIELD)
