+++ ssz/src/main/java/net/consensys/cava/ssz/annotation/package-info.java
@@ -1,0 +1,6 @@
+/**
+ * Simple Serialize (SSZ) annotations are used for automatic encoding/decoding routines handling
+ * standard Java class with annotations as SSZ model representation.
+ */
+
+package net.consensys.cava.ssz.annotation;
